LPD-33971 Extract AspectRatio and Resolution enums to use them in others classes
